# Conflicts:
#	ultralytics/cfg/datasets/CityEscapeYOLO.yaml
#	ultralytics/engine/validator.py
#	ultralytics/models/yolo/semseg/val.py
#	ultralytics/nn/tasks.py
#	ultralytics/utils/metrics.py
#	ultralytics/utils/plotting.py

+++ ultralytics/data/utils.py
@@ -182,7 +182,7 @@ def mask2polygon(mask, downsample_ratio=1):
     for c in contours:
         area = cv2.contourArea(c)
         if area > 10:
-            polygons.append(c[:, 0, :].astype(np.float_))
+            polygons.append(c[:, 0, :].astype(np.float32))
 
     if downsample_ratio != 1:
         polygons = polygons / downsample_ratio
